update nlu tests
