+++ tests/nlu/extractors/test_synonyms.py
