+++ tests/nlu/test_components.py
