+++ tests/nlu/utilities.py
@@ -1,6 +1,5 @@
 import tempfile
 
-import pytest
 import ruamel.yaml as yaml
 
 from rasa.nlu.config import RasaNLUModelConfig
@@ -9,11 +8,6 @@ from rasa.nlu.train import train
 
 
 def base_test_conf(pipeline_template):
-    # 'response_log': temp_log_file_dir(),
-    # 'port': 5022,
-    # "path": tempfile.mkdtemp(),
-    # "data": "./data/test/demo-rasa-small.json"
-
     return RasaNLUModelConfig({"pipeline": pipeline_template})
 
 
@@ -32,10 +26,6 @@ async def interpreter_for(component_builder, data, path, config):
     )
     interpreter = Interpreter.load(path, component_builder)
     return interpreter
-
-
-def temp_log_file_dir():
-    return tempfile.mkdtemp(suffix="_rasa_nlu_logs")
 
 
 class ResponseTest:
